update bevy to 0.19.1

+++ src/bevy/constants.rs
@@ -1,2 +1,2 @@
 pub const WINDOW_TITLE: &str = "Chess Engine";
-pub const WINDOW_SIZE: (f32, f32) = (1000., 700.);
+pub const WINDOW_SIZE: (u32, u32) = (1000, 700);
